'unhide_patterns' UCI option
For parameters that have the 'hidden' attribute set Transformer normally
generates a getter that returns a suitable default value for that parameter.
This getter overrides any getter implemented by the mapping.
This attribute is typically set on sensitive data like passwords; a get on
these parameters should always return e.g. an empty string.
However, sometimes the actual value is desired. With this UCI option it can
be configured for which parameters the 'hidden' attribute should be removed.