Expose the no-abort flag of the get() function.
Normally when an error is encountered while executing a get() the request is
aborted and only an error code and message is returned. While this is the
behavior mandated by TR-069 for a GetParameterValues it is less convenient
in other cases (webUI, TR-069 implementation retrieving the forced inform
parameters or notification values)..
Internally we recently added a no-abort flag that caused Transformer to keep
going as much as possible when an error is encountered. This changeset
exposes this flag to the outside.