Properly handle failures to send a response.
The original code was written under the assumption that the socket was a
blocking one. At a certain moment the default behavior for a socket created
via the lua-tch library (accidentally) changed to be non-blocking. This went
unnoticed for several years until a culmination of speed improvements in
Transformer, a bigger datamodel and a test ACS on a slow CI machine that
couldn't keep up caused the sendto() to fail regularly and Transformer
dropping the response packet. This manifested itself in cwmpd as either
missing data or a socket read timing out (+ an error trace).

We now check for a WOULDBLOCK return code from sendto() and wait for up to
15 seconds for the socket to become writable again so we can resend the
response or drop it (with an error logged).