Return instances in numeric order of instance number.
Instances that are addressed using instance numbers are now returned in
numerical order. Internally the instance identifiers are strings because
we also support name-based addressing of instances. Sorting strings and
expecting numbers in those strings to be sorted 'naturally' isn't trivial
so we've taking a shortcut to keep it efficient. Because the focus is on
instance numbers and we know that in that case the instance identifier
string contains only numbers we can sort things by sorting short strings
before long ones and strings of equal length alphabetically.
Disadvantage is that for name-based instances they are sorted short names
before long ones and names of equal length alphabetically. But at least
they are now returned in a deterministic order.
Note that this commit changes the structure of the first table returned
by Persistency:sync() (and thus also that of Typestore:synchronize()).
That table now has an array part with the sorted irefs. This means you
can no longer iterate over this table using pairs(); instead use ipairs().