Rework the tch_async library.
The original tch_async code had a race condition and no way to stop scripts
after a timeout. These problems could cause the async loop to stop.

The original async queue code was cleaned up, restructured and moved to async.c
It will now log errors to syslog, both for internal failures and for commands
failing to execute correctly.
The timeout used for executing commands is a property of the queue and is adjustable.
The default timeout for executing commands is 30 seconds.

The original limit on queue size was removed.

Executing the commands is no longer done with the system() call as that does not
allow to specify a timeout. The code for executing the command is in execute.c.
The code for execute() forks twice. First into the process that will implement
the timeout, then to the process that will finally exec() the given command.

Finally the tch_async.c code now only contains the Lua interface.
It will create a single global queue and expose execute(), stats() and timeout()

The api for execute was changed so that it now returns a more logical boolean
value instead of an integer. For the current usage in transformer this does not
matter as the result of execute() is ignored.